refactor: N+1 문제 해결 (#114)
