+++ src/main/java/ku_rum/backend/domain/oauth/application/CustomOAuth2UserService.java
@@ -27,7 +27,6 @@ import java.util.Map;
 public class CustomOAuth2UserService implements OAuth2UserService<OAuth2UserRequest, OAuth2User> {
 
     private final UserRepository userRepository;
-//    private final JwtTokenProvider jwtTokenProvider;
 
     public CustomOAuth2UserService(UserRepository userRepository) {
         this.userRepository = userRepository;
