+++ src/main/java/ku_rum/backend/domain/recruitment/presentation/RecruitmentController.java
@@ -4,11 +4,10 @@ import ku_rum.backend.domain.recruitment.application.RecruitmentService;
 import ku_rum.backend.global.response.BaseResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import static ku_rum.backend.domain.notice.presentation.CrawlingResponse.START_CRAWLING;
+import static ku_rum.backend.domain.notice.dto.response.CrawlingResponse.START_CRAWLING;
 
 @RestController
 @RequestMapping("/api/v1/recruitments")
